refactor: move timeline components into its subfolder
