+++ frontend/src/components/matchEvents/MatchDetailsContent.tsx
@@ -4,7 +4,7 @@ import { useMatchEvents } from '@/hooks/useMatchEvents';
 import ErrorState from '../feedback/ErrorState';
 import MatchHeader from './header/MatchHeader';
 import MatchHeaderSkeleton from './header/MatchHeaderSkeleton';
-import MatchTimeline from './MatchTimeline';
+import MatchTimeline from './timeline/MatchTimeline';
 
 interface MatchDetailsProps {
   matchId: number;
