+++ frontend/src/components/matchEvents/timeline/EventCard.tsx
@@ -4,7 +4,7 @@ import { ICON_SIZE } from '@/constants/matchEvents';
 
 import { cn } from '@/lib/utils';
 
-import { Card } from '../ui/card';
+import { Card } from '../../ui/card';
 
 import {
   formatEventMinute,
